- This checkin removes a lot of warning messages when compiled with -Wall.   There are still a couple of warning messages that I couldn't remove yet. - The process revealed a bug in FileConnection restoration. The fix is included   in this checkin. - Another bugs (that never triggered) involved the precedence of binary '&' and   '!='/'=='.

+++ trunk/dmtcp/src/sysvipc.cpp
@@ -200,7 +200,7 @@ void dmtcp::SysVIPC::removeStaleShmObjects()
       staleShmids.push_back(i->first);
     }
   }
-  for (int j = 0; j < staleShmids.size(); ++j) {
+  for (size_t j = 0; j < staleShmids.size(); ++j) {
     _shm.erase(staleShmids[j]);
   }
 }
